requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
python-dotenv
pypdf
python-multipart
pydantic